Moved 'real' code out of headers
DO NOT UNDO THIS. Putting method bodies in headers is very bad practice, to
the point where I've heard it's not even allowed in C (as opposed to C++).
I understand there's now three .cpp files to compile instead of one, but
chains of includes are NOT A SUBSTITUTE for a proper build system. If you were
on Linux, you could simply go to the root of the repository and type 'make
renderer' and it would compile everything. (It actually doesn't, because
circuit-render-error.h has gone missing, but that can be cleared up.)

As for the directory structure of the repository, we can have this argument if
you like, but I guarantee you most of the team will take my side. This project
is going to get way too big to not organise the source code.