Simplified renderer API
Rather than having a bunch of trivial getter functions and and a preprocessor macro, I created one function [void* crGetFrame(uint, uint)], which takes the required width and height, calls crDraw with them, and returns a pointer to the raw image data. The Qt code will handle conversion to QImage on its end.